requirements.txt
numpy==1.23.2
absl-py==1.2.0
albumentations==1.2.1
anyio 
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens
astunparse==1.6.3
attrs==22.1.0
Babel 
backcall 
beautifulsoup4
bleach==5.0.1
brotlipy
cachetools==5.2.0
certifi 
cffi 
charset-normalizer 
click==8.1.3
coloredlogs==15.0.1
cryptography
cycler==0.11.0
Cython==0.29.31
debugpy 
decorator 
deepface==0.0.75
defusedxml 
dlib==19.24.0
docopt==0.6.2
dotmap==1.3.30
easydict==1.9
entrypoints
executing
fastjsonschema==2.16.1
filelock==3.7.1
fire==0.4.0
Flask==2.1.3
flatbuffers==1.12
fonttools==4.34.4
gast==0.4.0
gdown==4.5.1
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
humanfriendly==10.0
idna 
imageio==2.20.0
importlib-metadata==4.12.0
imutils==0.5.4
insightface==0.6.2
install==1.3.5
ipykernel 
ipython 
ipython-genutils 
ipywidgets==8.0.1
itsdangerous==2.1.2
jedi 
Jinja2==3.1.2
joblib==1.1.0
json5 
jsonschema==4.14.0
jupyter==1.0.0
jupyter-client 
jupyter-console==6.4.4
jupyter-contrib-core==0.4.0
jupyter-contrib-nbextensions==0.5.1
jupyter-core 
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.5.0
jupyter-server 
jupyterlab 
jupyterlab-pygments==0.2.2
jupyterlab-server 
jupyterlab-widgets==3.0.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.1
llvmlite==0.39.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe 
matplotlib==3.5.2
matplotlib-inline 
mediapipe==0.8.10.1
mistune 
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.2.1
mtcnn==0.1.1
nb-conda==2.2.1
nb-conda-kernels 
nbclassic 
nbclient 
nbconvert 
nbformat==5.4.0
nest-asyncio 
networkx==2.8.5
notebook 
numba==0.56.3
oauthlib==3.2.0
olefile 
onnx==1.12.0
onnxruntime==1.12.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging 
pandas==1.4.3
pandocfilters 
parso 
pexpect 
pickleshare 
Pillow==9.2.0
pipreqs==0.4.11
prettytable==3.3.0
prometheus-client 
prompt-toolkit 
protobuf==3.19.4
ptyprocess 
pure-eval 
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser 
Pygments 
pynndescent==0.5.7
pyOpenSSL 
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks 
python-dateutil 
pytz 
PyWavelets==1.3.0
PyYAML==6.0
pyzmq 
qtconsole==5.3.2
QtPy==2.2.0
qudida==0.0.4
requests 
requests-oauthlib==1.3.1
retina-face==0.0.12
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
Send2Trash 
sip==4.19.13
six 
sniffio 
soupsieve==2.3.2.post1
stack-data 
sympy==1.10.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
testpath 
thop==0.1.1.post2207130030
threadpoolctl==3.1.0
tifffile==2022.5.4
tinycss2==1.1.1
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
tornado 
tqdm==4.64.0
traitlets==5.3.0
typing_extensions 
umap-learn==0.5.3
urllib3 
utils==1.0.1
wcwidth 
webencodings==0.5.1
websocket-client 
Werkzeug==2.2.1
widgetsnbextension==4.0.2
wrapt==1.14.1
yarg==0.1.9
zipp==3.8.1
